updating everything, combining things from all of us,and seeing if it works!
