+++ requirements.txt
@@ -6,3 +6,7 @@ matplotlib==3.5.1
 prophet==1.1.1
 scikit-learn == 1.0.2
 tensorflow ==2.11.0
+libta-lib ==0.4.0
+ta-lib == 0.4.19
+finta ==1.3
+ta == 0.10.2
